+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [7.042]
 
-Deps updated
+* Deps updated
+* added fn 'map-column->columns' ([#178])(https://github.com/scicloj/tablecloth/issues/178)
 
 ## [7.029]